feat: add linux installation

+++ karabiner/setup.sh
@@ -2,11 +2,11 @@
 source "$HOME/dotconfig/setup-utils.sh"
 
 action="install karabiner"
-if ! check_dependency karabiner-elements && confirm "$action"; then
+if is_mac && ! check_dependency karabiner-elements && confirm "$action"; then
   brew install --cask karabiner-elements
 fi
 
 action="link karabiner config"
-if confirm "$action"; then
+if is_mac && confirm "$action"; then
   link_config "$HOME/dotconfig/karabiner/karabiner.json" "$HOME/.config/karabiner/karabiner.json"
 fi
